Move golden tests of `qr_code_scanner` into the `test_goldens` folder
