+++ lib/qr_code_scanner/test_goldens/flutter_test_config.dart
