+++ lib/qr_code_scanner/test_goldens/goldens/scanner_page.png
