+++ lib/qr_code_scanner/test_goldens/overlay/goldens/scan_overlay.phone.png
