+++ lib/qr_code_scanner/test_goldens/overlay/goldens/scan_overlay.tablet_landscape.png
