+++ lib/qr_code_scanner/test_goldens/overlay/scan_overlay_test.dart
